Avoid direct access to (struct vcc).sb
Instead there are two funtions vcc_Complain() and vcc_Complainf() for
either constant or formatted error messages. In the process the `sb`
field was renamed to `err_sb`.

There was one use of `err_sb` for file coordinates that couldn't be
avoided because it might also be used on an other VSB. An indirection
was added to deal with that.

This is a semi-automated change thanks to Coccinelle and sed.