Tie workspace logs to their transactions
That is, when it's a request or backend transaction, otherwise it's
still logged with VXID 0. The goal is to allow a correlation between
VCL statements and their workspace footprint by sharing the same
transaction instead of being split between raw and vxid scopes. That
kind of debugging becomes possible when the workspace debug flag is
used and VCL_trace records are not masked.

Given the verbosity of both workspace and trace logs, we would need
new tools to process transactions for this purpose and until we can
link everything we have a chicken-egg problem.

This change introduces a new thread-local variable to keep track of a
worker's VSL when it is given one. Not all workspace operations from
a transaction are logged, the missing bits are early initialization
and teardown: in other words before and after VCL execution, not VCL
statements that could be debugged using such logs.

Grabbing the VSL from the thread-local req or bo isn't possible, the
VSL may not be ready to use. We could introduce a VSL_IsReady() or
some other means to find whether we can use it, but I felt that we
had a clear signal in the code regarding when we can start doing that
via the worker.

This introduces a dependency between workspace and workers, so the
latter need to be initialized earlier.