build: Only --enable-developer-warnings adds warnings
This moves the definition for --enable-developer-warnings up, before
warnings are added. The wflags.py script is the first one to feed
DEVELOPER_CFLAGS, which allows to remove redundant -Wall and -Werror
flags. Warnings that were added directly to CFLAGS are now limited
to DEVELOPER_CFLAGS, but when we --enable-developer-warnings they
will join the OCFLAGS to trickle down to distcheck.

The result is a simpler cc_command for production, free of warnings,
and limited to actual CFLAGS injected at configure time by packaging
systems or sysadmins.